pyproject.toml
[project]
name = "resume_builder"
version = "0.1.0"
description = "AI Resume Builder using CrewAI and Streamlit"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.134.0,<1.0.0",
    "pysqlite3-binary == 0.5.4",
    "streamlit>=1.28.0",
    "reportlab>=3.6.0",
    "python-docx>=0.8.11",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "tenacity>=8.0.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.0.0"
]

[project.scripts]
resume_builder = "resume_builder.main:run"
run_crew = "resume_builder.main:run"
train = "resume_builder.main:train"
replay = "resume_builder.main:replay"
test = "resume_builder.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[tool.hatch.build.targets.wheel]
packages = ["src/resume_builder"]